Fix build issues: Java 21 setup, test assertions, and Double.MIN_VALUE bug

+++ jena-falkordb-adapter/src/test/java/com/falkordb/jena/query/GeoSPARQLToCypherTranslatorTest.java
@@ -102,8 +102,8 @@ public class GeoSPARQLToCypherTranslatorTest {
         
         assertNotNull(result);
         // Center point: lat (51.50+51.52)/2 = 51.51, lon (-0.15+-0.10)/2 = -0.125
-        assertEquals(51.51, params.get("rect_lat"), 0.001);
-        assertEquals(-0.125, params.get("rect_lon"), 0.001);
+        assertEquals(51.51, (Double) params.get("rect_lat"), 0.001);
+        assertEquals(-0.125, (Double) params.get("rect_lon"), 0.001);
         // Check bounding box parameters
         assertEquals(51.50, params.get("rect_minLat"));
         assertEquals(51.52, params.get("rect_maxLat"));
@@ -381,8 +381,8 @@ public class GeoSPARQLToCypherTranslatorTest {
         assertNotNull(result);
         // Bounding box: lat [48.8566, 52.5200], lon [-0.1278, 13.4050]
         // Center: lat = 50.6883, lon = 6.6386
-        assertEquals(50.6883, params.get("route_lat"), 0.001);
-        assertEquals(6.6386, params.get("route_lon"), 0.001);
+        assertEquals(50.6883, (Double) params.get("route_lat"), 0.001);
+        assertEquals(6.6386, (Double) params.get("route_lon"), 0.001);
         assertEquals(48.8566, params.get("route_minLat"));
         assertEquals(52.5200, params.get("route_maxLat"));
         assertEquals(-0.1278, params.get("route_minLon"));
@@ -481,8 +481,8 @@ public class GeoSPARQLToCypherTranslatorTest {
         
         assertNotNull(result);
         // Bounding box: lat [48.8566, 52.5200], lon [-0.1278, 13.4050]
-        assertEquals(50.6883, params.get("cities_lat"), 0.001);
-        assertEquals(6.6386, params.get("cities_lon"), 0.001);
+        assertEquals(50.6883, (Double) params.get("cities_lat"), 0.001);
+        assertEquals(6.6386, (Double) params.get("cities_lon"), 0.001);
         assertEquals(48.8566, params.get("cities_minLat"));
         assertEquals(52.5200, params.get("cities_maxLat"));
         assertEquals(-0.1278, params.get("cities_minLon"));
